RTF-1378 - change of return type, from int to long, of Run and RunAsync methods of IImagingJobService

+++ source/Relativity.Testing.Framework.Api/Strategies/Imaging/Job/IImagingJobRunStrategy.cs
@@ -5,8 +5,8 @@ namespace Relativity.Testing.Framework.Api.Strategies
 {
 	internal interface IImagingJobRunStrategy
 	{
-		int Run(int workspaceId, int imagingSetId, ImagingSetJobRequest imagingSetJobRequest = null);
+		long Run(int workspaceId, int imagingSetId, ImagingSetJobRequest imagingSetJobRequest = null);
 
-		Task<int> RunAsync(int workspaceId, int imagingSetId, ImagingSetJobRequest imagingSetJobRequest = null);
+		Task<long> RunAsync(int workspaceId, int imagingSetId, ImagingSetJobRequest imagingSetJobRequest = null);
 	}
 }
